Server/requirements.txt
transformers==4.18.0
torch>=1.10.0
torchaudio>=0.9.0
flask>=2.0.3
gunicorn>=20.1.0
python-Levenshtein

#transformers==4.18.0
#
#ptvsd==4.3.2 # Required for debugging.
#torchaudio>=0.9.0
#pytorch==1.12.0
#requests>=2.27.1
